[session] ModFactory: Add exception to refresh waitinfo screen (WAB-9164)
Message can change in waitinfo module and sometime, the
transition:module send is not processed because another module:waitinfo
immediately follows in the same sesman socket 'read'.

This added condition allows to reload waitinfo module.

May be we should add a force_refresh option so that sesman can force
to refresh a already loaded module ?